Force standalone-argp to use fputc instead of putc under Win32.
As the patch indicates, 'putc' is doing something funny on Win32
which causes it to abort at runtime.  Given that the parameter that
the standalone argp passes to it is 'stream,' which is probably
defined somewhere in Mingw as a macro, and putc allows for evaluation
of parameters, it's probably doing something screwy.  We now use
'fputc' in this case, and everyone is suddenly happy and everything
works as we expect.