Eek!  UniConfRoot was pulling in *all* the generators by default.
uniconfroot.cc used WV_LINK_TO(UniGenHack), which means any
statically-linked UniConf-using program pulled in *all* the generators,
including the useless ones, the SSL ones, and so on, and became gigantic.

The purpose of WV_LINK_TO() is to let you link to what you need.  If you
want all of them, you can WV_LINK_TO(UniGenHack).  But if you don't, we
shouldn't force it on you.