WvError: set*() functions now return const WvError&.
This makes it more convenient to do things like:
	return WvError().set(ENOENT)
in functions that want to return a WvError.

We could add some constructors to save extra typing (so "return ENOENT"
would work) but I'm not sure if that's a good idea or not.

Note that this doesn't change the ABI, since the return value used to be
void (thus ignored).