urlget/wvhttpstream.cc:  Fix 'connection refused' running on Win32.
... when pipelining is enabled.  First of all, it's not really
connection refused, it's actually connection aborted, but we seem
to be translating the error because of some Wine bug from long ago?

Basically, if the pipelining error got a 404 (pretty probable), the
server we were connecting to would shut the connection down
immediately, which would return EOF on Linux, but in Windows, the
dreaded 10053.

The dying attempt at pipelining would then go ahead and close your
*real* request because it died with an error.

The 'fix' here is to ignore errors from failed pipelining requests,
since we really don't trust them anyway (they failed), and so
we modify WvHttpStream::close() to prevent propagation of errors
when a pipelining failure occurs.