WvError: add an onerror() callback.
Now you can set a function to be called whenever an error is set.  If an
error was *already* set on this WvError object, the callback doesn't get
called (because it isn't going to change the WvError state anyway).  If you
want, your callback can call reset().

This is handy for making *sure* a message will be printed as soon as an
error occurs.  I guess it would work on WvStream, too, since it inherits
from WvErrorBase, though I haven't tried that yet.