Makefile.in:  Actually allow -jN building.
Despite other attempts, the real problem to building WvStreams in
parallel was that we were explicitly adding LIBS variables which
were used by the linker, but not specified as build dependencies.
This meant the linker was often invoked too early.

Finally, with some clever and some mundane work, this is fixed.

All libraries we build are listed as build dependencies, so Make
can build them in the correct order.