wvstreams/win32: never send() more than 64k to a socket at once.
It seems Windows (but not Wine) can buffer a huge amount of data if you try
to write too much at once, which means programs get confused about how long
it's taking to write "one block."

Among other things, this confused WvHttpStream, which had a per-block
timeout of 2 minutes.  For 64k, a 2 minute timeout is pretty reasonable, but
when uploading an 8 meg file, it all ended up in one block. 8 megs in 2
minutes *isn't* so reasonable.

This solves spurious "client timed out" problems in EQL Data, for example.