wvbuild/Makefile: get rid of spurious $3 that caused parallel make warnings.
Somehow, the $3 parameter to $(make_subdir) was getting substituted as -j1.
How?  I have no idea.  I can't find -j1 anywhere in the code, and certainly
nobody in wvbuild/Makefile is passing a third parameter to make_subdir.

This is insane and probably a bug in make - but I don't care enough to hunt
it down.  Since nobody ever used the third parameter, whatever it might be
for, let's just not refer to it.  Problem solved, and parallel makes work
reliably again.  Super.