uniregistrygen: support REG_EXPAND_SZ, ensure there's a null terminator.
Grr, in Windows Vista and above, some keys that used to be REG_SZ are not
REG_EXPAND_SZ, which it turns out uniregistrygen didn't support... and
silently returned null instead.

So now we add support for that type (just return it as if it were REG_SZ).

While I was reading the docs for RegQueryValueEx, I saw that some keys are
also not necessarily null terminated, so I fixed that up too.

Furthermore, we now print to stderr if we ever find a registry key where we
don't know how to read data of that type.