Add -j12 to configure and configure-mingw scripts.
They're kind of misnamed, because they "configure" wvstreams but only
after building all of wvports, which takes a really long time if you
don't have any parallelism.

This isn't a great fix (would be better to have more of the build
powered directly by make dependencies), but it'll work for now.