wvbuild: replace top-level makefiles with redo scripts.
This improves dependency handling quite a bit, though there's still
plenty of cleaning to do, and more Makefiles to remove.

We currently have a bit of "special" handling for the multi-step
wvstreams autogen-configure-make build process.  If we migrate
wvstreams to redo also, we could remove that stuff from this layer.