wvports/*/Makefile: when calling $(MAKE), prefix with +.
This will be a short-lived fix since I'm converting to redo, but it's a
fairly serious error.  Because $(MAKE) is in a macro, its presence
doesn't automatically cause a + to be prepended to calls to $(compile)
in wvports.mk.  So each of the wvports was being compiled with the
equivalent of -j1.  This makes it act correctly.